+++ src/kitconcept/glossary/tests/test_term.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from kitconcept.glossary.interfaces import ITerm
+from kitconcept.glossary.interfaces import IGlossaryTerm
 from kitconcept.glossary.testing import INTEGRATION_TESTING
 from plone import api
 from plone.dexterity.interfaces import IDexterityFTI
@@ -19,27 +19,27 @@ class GlossaryTypeTestCase(unittest.TestCase):
         with api.env.adopt_roles(['Manager']):
             self.g1 = api.content.create(self.portal, 'Glossary', 'g1')
 
-        self.t1 = api.content.create(self.g1, 'Term', 't1')
+        self.t1 = api.content.create(self.g1, 'GlossaryTerm', 't1')
 
     def test_adding(self):
-        self.assertTrue(ITerm.providedBy(self.t1))
+        self.assertTrue(IGlossaryTerm.providedBy(self.t1))
 
     def test_adding_outside_glossary(self):
         from plone.api.exc import InvalidParameterError
         with self.assertRaises(InvalidParameterError):
-            api.content.create(self.portal, 'Term', 'test')
+            api.content.create(self.portal, 'GlossaryTerm', 'test')
 
     def test_fti(self):
-        fti = queryUtility(IDexterityFTI, name='Term')
+        fti = queryUtility(IDexterityFTI, name='GlossaryTerm')
         self.assertIsNotNone(fti)
 
     def test_schema(self):
-        fti = queryUtility(IDexterityFTI, name='Term')
+        fti = queryUtility(IDexterityFTI, name='GlossaryTerm')
         schema = fti.lookupSchema()
-        self.assertEqual(ITerm, schema)
+        self.assertEqual(IGlossaryTerm, schema)
 
     def test_factory(self):
-        fti = queryUtility(IDexterityFTI, name='Term')
+        fti = queryUtility(IDexterityFTI, name='GlossaryTerm')
         factory = fti.factory
         new_object = createObject(factory)
-        self.assertTrue(ITerm.providedBy(new_object))
+        self.assertTrue(IGlossaryTerm.providedBy(new_object))